Implement protocol chooser
I played a little fast and loose with this one, I copied it over
mostly verbatim when it could use a bit of a rewrite. The main
thing that bothers me is the way protocol-chooser writes directly
to the ShinyServer global object, and also that it does so much
work directly upon being require()'d instead of needing to be
explicitly called with an exported function.

Note that the list of allowed protocols from the server no longer
is passed by template rendering, but as an init option.