Replace _ => {} with () => {}
Using _ is not actually semantically equivalent to function().
I thought _ was a magic symbol like in Go but nope, it's just
a regular argument. In the mocha tests, I had noticed that
tests were acting like they were async when I replaced
function() with _ =>, it's because the arity of the function
is changing.