send a dummy protocol instead of an empty whitelist
SockJS regards an empty `protocols_whitelist` option as an indication that all
protocols should be enabled. To avoid this behavior, pass a "dummy" protocol to
SockJS when we end up with an empty set of whitelisted protocols. This isn't
useful in the real world but helps validate behavior when all protocols end up
disabled through whatever technique.